refactor(view): extract shader spawning logic into separate module
